% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenODS.R
\name{ods_api}
\alias{ods_api}
\title{Open ODS}
\usage{
ods_api(path, query = NULL)
}
\arguments{
\item{path}{String: Path for query.}

\item{query}{String: Query parameter.}
}
\value{
Object of class \code{openods}, a list with three elements:
          \emph{content} The parsed content returned by the API
          \emph{path} The path used by \code{GET}
          \emph{response} The raw JSON data returned by the API
}
\description{
Organisational API for the NHS
}
\details{
See \url{http://www.openods.co.uk}
}
\examples{
ods_api("api/organisations", query = "RADCLIFFE")
}
