% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenODS.R
\name{ods_organisations}
\alias{ods_organisations}
\title{Return organizations from a search on any of their characteristics}
\usage{
ods_organisations(query = NULL)
}
\arguments{
\item{query}{String.}
}
\value{
openods object.
}
\description{
Return organizations from a search on any of their characteristics
}
\examples{
ods_organisations(query = "q=RADCLIFFE")
}
