% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenODS.R
\name{ods_rolecode}
\alias{ods_rolecode}
\title{Return all organizations with the same role code}
\usage{
ods_rolecode(code = NULL, raw = TRUE)
}
\arguments{
\item{code}{String: the ODS role code.}

\item{raw}{Logical: whether to return the raw data}
}
\value{
openods object, a list with three components
}
\description{
Return all organizations with the same role code
}
\examples{
ods_rolecode(code = "RO101")
}
