% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenODS.R
\name{ods_roles}
\alias{ods_roles}
\title{Retrieve role types}
\usage{
ods_roles(code = NULL)
}
\arguments{
\item{code}{String: if \code{code = NULL} (the default), will return a tibble with all available role
codes and their names.  If \code{code} is a valid role type code, will return information about
that role.}
}
\value{
A tibble with two columns, code and name
}
\description{
Retrieve role types
}
\examples{
ods_roles()
ods_roles("RO0101")
}
