% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenODS.R
\name{print.openods}
\alias{print.openods}
\title{Print method for openods class}
\usage{
\method{print}{openods}(x, ...)
}
\arguments{
\item{x}{An \code{openods} object}

\item{...}{}
}
\value{
Nothing
}
\description{
Print method for openods class
}
\examples{
print(x)
}
